Extract rendering module from AgentCanvas (74.8% reduction)
MAJOR MILESTONE: Extracted all drawing logic into dedicated renderer module

Created agent-canvas-renderer.js:
- 2,812 lines of pure rendering code
- 31 drawing methods extracted
- AgentCanvasRenderer class with clean interface
- Constructor: (ctx, state, canvas, parent)
- All drawing logic preserved exactly
- Comprehensive JSDoc documentation

Methods extracted:
- Connection rendering (9): drawConnections, drawArrow, drawChainConnections, etc.
- Task & agent rendering (2): drawTaskFlows, drawAgents
- Panel rendering (6): drawExpandedTaskPanel, drawExpandedAgentPanel, etc.
- Button rendering (5): drawCreateTaskButton, drawAddAgentButton, etc.
- UI elements (3): drawNotifications, drawContextMenu, drawHelpOverlay
- Combiner rendering (2): drawCombinerNodes, drawPort
- Helper methods (2): roundRect, wrapText

Updated agent-canvas.js:
- Reduced from 6,432 → 1,622 lines (74.8% reduction!)
- Main draw() method now orchestrates via this.renderer.*
- All drawing method implementations removed
- Preserved: state management, event handling, business logic
- Renderer instantiated in constructor

Architecture improvements:
- Clean separation of concerns
- Rendering isolated from application logic
- State accessed via this.state (dependency injection)
- All rendering in single, focused module

Impact:
- agent-canvas.js: 6,432 → 1,622 lines (-4,810 lines, -74.8%)
- agent-canvas-renderer.js: 2,812 lines (new)
- Total code better organized, more maintainable

Related tasks: #5.0, #6.0
